# KAFKA-9820: validateMessagesAndAssignOffsetsCompressed allocates unused iterator