--- a/jmh-benchmarks/src/main/java/org/apache/kafka/jmh/record/RecordBatchIterationBenchmark.java
+++ b/jmh-benchmarks/src/main/java/org/apache/kafka/jmh/record/RecordBatchIterationBenchmark.java
@@ -98,8 +98,8 @@ public class RecordBatchIterationBenchmark {
 
     @Setup
     public void init() {
-        // zero starting offset is much faster for v1 batches, but that will almost never happen
-        // for v2 batches we use starting offset = 0 these batches are relative to the base
+        // For v0 batches a zero starting offset is much faster but that will almost never happen.
+        // For v2 batches we use starting offset = 0 as these batches are relative to the base
         // offset, and measureValidation will mutate these batches between iterations
         startingOffset = messageVersion == 2 ? 0 : 42;